+++ app/javascript/components/containers/communication/messages/ShowMessage.jsx
@@ -20,6 +20,9 @@ class ShowMessage extends React.Component {
     }
 
     onDeleteClick(event) {
+        let shouldDelete = confirm('Are you sure?');
+        if (!shouldDelete) return;
+
         MessageApi.destroy(this.props.match.params.id).then(() => {
             history.push('/communication/messages');
         });
