+++ app/javascript/components/containers/communication/sign-ups/EditSignUp.jsx
@@ -20,6 +20,9 @@ class EditSignUp extends React.Component {
     }
 
     onDeleteClick(event) {
+        let shouldDelete = confirm('Are you sure?');
+        if (!shouldDelete) return;
+
         ContactApi.destroy(this.props.match.params.id).then(() => {
             history.push('/communication/sign-ups');
         });
