+++ app/javascript/components/containers/communication/sign-ups/ShowSignUp.jsx
@@ -21,6 +21,9 @@ class ShowSignUp extends React.Component {
     }
 
     onDeleteClick(event) {
+        let shouldDelete = confirm('Are you sure?');
+        if (!shouldDelete) return;
+
         ContactApi.destroy(this.props.match.params.id).then(() => {
             history.push('/communication/sign-ups');
         });
