+++ app/javascript/components/containers/communication/volunteers/ShowVolunteer.jsx
@@ -21,6 +21,9 @@ class ShowVolunteer extends React.Component {
     }
 
     onDeleteClick(event) {
+        let shouldDelete = confirm('Are you sure?');
+        if (!shouldDelete) return;
+
         VolunteerApi.destroy(this.props.match.params.id).then(() => {
             history.push('/communication/volunteers');
         });
